requirements.txt
discord.py = latest
python-dotenv = latest